refactor : 이미지 타입 수정
