+++ src/main/java/com/smile/petpat/common/response/config/CustomClientHttpRequestInterceptor.java
@@ -1,4 +1,4 @@
-package com.smile.petpat.config;
+package com.smile.petpat.common.response.config;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
