+++ src/main/java/com/smile/petpat/common/response/config/JpaConfig.java
@@ -1,4 +1,4 @@
-package com.smile.petpat.config;
+package com.smile.petpat.common.response.config;
 
 import org.springframework.context.annotation.Configuration;
 import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
