+++ src/main/java/com/smile/petpat/common/response/config/comm/Timestamped.java
@@ -1,4 +1,4 @@
-package com.smile.petpat.config.comm;
+package com.smile.petpat.common.response.config.comm;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
